Fix int overflow in DeviceTensor capacity for large-scale PA (>8.8M hex elements)

+++ fem/integ/bilininteg_diffusion_pa.cpp
@@ -136,7 +136,7 @@ void DiffusionIntegrator::AssemblePA(const FiniteElementSpace &fes)
    symmetric = (coeff_dim != dims*dims);
    const int pa_size = symmetric ? symmDims : dims*dims;
 
-   pa_data.SetSize(pa_size * nq * ne, mt);
+   pa_data.SetSize((int64_t)pa_size * nq * ne, mt);
    internal::PADiffusionSetup(dim, sdim, dofs1D, quad1D, coeff_dim, ne,
                               ir->GetWeights(), geom->J, coeff, pa_data);
 }
